refactor(upload): rename and move for better hierarchy
