+++ src/admin/upload/components/AddWatermarkButton/canvas.js
