+++ src/admin/upload/components/Loader/style.scss
